+++ src/cores/M48T08/M48T08.hpp
@@ -34,7 +34,7 @@ public:
     M48T08();
     ~M48T08();
 
-    void IncrementClock(const uint32_t ns);
+    void IncrementClock(const size_t ns);
 
     uint8_t GetByte(const uint16_t addr) const;
     void SetByte(const uint16_t addr, const uint8_t data);
